\name{rAutoARR-package}
\alias{rAutoARR-package}
\alias{rAutoARR}
\docType{package}
\title{
\packageTitle{rAutoARR}
}
\description{
\packageDescription{rAutoARR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rAutoARR}
\packageIndices{rAutoARR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rAutoARR}

Maintainer: \packageMaintainer{rAutoARR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
